fixed Books chart filter
