+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "iare",
-  "version": "0.3.0.3",
-  "version_name": "0.3.0.3",
+  "version": "0.3.0.4",
+  "version_name": "0.3.0.4",
   "private": true,
   "type": "module",
   "dependencies": {
